requirements.txt
# CarnavalPlay — Dependencias Python
# Instalar con: pip install -r requirements.txt

# ── Framework web ──────────────────────────────────────────────────
Flask==3.0.3
flask-socketio==5.3.6
python-engineio==4.9.1
python-socketio==5.11.3

# ── Autenticación ──────────────────────────────────────────────────
flask-login==0.6.3
bcrypt==4.1.3

# ── Base de datos ──────────────────────────────────────────────────
SQLAlchemy==2.0.30

# ── HTTP / APIs ────────────────────────────────────────────────────
requests==2.32.3
google-api-python-client==2.134.0
google-auth==2.30.0

# ── Scraping sin cuota de API (fallback yt-dlp) ────────────────────
yt-dlp>=2024.1.1

# ── Scheduler ──────────────────────────────────────────────────────
APScheduler==3.10.4

# ── Config ─────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Servidor WSGI para Windows (no gunicorn) ───────────────────────
waitress==3.0.0

# ── IA / Groq (chatbot) ────────────────────────────────────────────
groq==0.9.0

# ── Utilidades ─────────────────────────────────────────────────────
Pillow==10.3.0
